Move serial port loggers/helpers to subpackage serial_helpers
